requirements.txt
Flask==2.0.1
flask-mysqldb==0.2.0
mysql-connector-python==8.0.26
Werkzeug==2.0.1
flask-cors==3.0.10